+++ tests/test_tutorial_priors.py
@@ -26,11 +26,7 @@ from petab.v2 import Problem
 from petab.v2.lint import lint_problem
 
 from pybnf.petab import import_job
-from pybnf.petab.bngl_model import register_bngl
 from pybnf.petab.parameters import read_parameter_table, free_parameter_from_row
-
-# Teach petab to load `language: bngl` problems (idempotent).
-register_bngl()
 
 _REPO_ROOT = Path(__file__).resolve().parents[1]
 _LESSON = _REPO_ROOT / 'examples' / 'tutorial' / '15_petab_priors'